Fix how wide runes are displayed
Internal cell buffer represent each rune which should be displayed.
This cell buffer is translated into tcell representation in `draw`.
If wide rune is presented, it will be represented in `tcell` by one cell
with the rune and following with "empty" cells.

Internal representation is fixed in multiple places so it's not mixed
with interpreting rune width.

This fixes also problem with `SetWritePos` function.